Do not infer bigint by inspecting the object but rely on hints
